add function sent email, modify tasks

+++ service/operation/views.py
@@ -3,7 +3,7 @@ import json
 from django.shortcuts import render, redirect
 from operation.forms import FileUploadForm
 from operation.models import FileModel
-from operation.tasks import log_handler
+from operation.tasks import read_logs
 
 
 def upload_file(request):
@@ -14,8 +14,8 @@ def upload_file(request):
             file.user = request.user
             file.file_name = request.FILES['file_path'].name
             file.save()
-            file_id = file.id
-            log_handler.delay(file_id)
+
+            read_logs.delay(file.id)
 
             return redirect("succes")
 
@@ -34,8 +34,19 @@ def uploaded_files(request):
 def delete_file(request, file_id):
     try:
         file = FileModel.objects.get(id=file_id, user=request.user)
-        file.file_path.delete()
+        file_path = file.file_path.path
+
+        with open("logs.json", "r") as json_file:
+            json_data = json.load(json_file)
+            json_data = [item for item in json_data if file_path not in item]
+
+        with open("logs.json", "w") as json_file:
+            json.dump(json_data, json_file)
+
+        if file_path:
+            file.file_path.delete()
         file.delete()
+
     except FileModel.DoesNotExist:
         pass
 
@@ -51,12 +62,12 @@ def update_file(request, file_id):
             file.file_path.delete(save=False)
 
             file.file_path = request.FILES['file_path']
-            file.file_name = request.FILES['file_path']
+            file.file_name = request.FILES['file_path'].name
             file.is_new = False
             file.is_changed = True
 
             file.save()
-            log_handler.delay(file.id)
+            read_logs.delay(file.id)
 
         return redirect("succes")
     else:
@@ -71,16 +82,14 @@ def succes_view(request):
 
 def report_info(request, file_id):
     file = FileModel.objects.get(pk=file_id)
-    file_path = file.file_path.path
-    file_name = file.file_name
 
     with open("logs.json", "r") as json_file:
         json_data = json.load(json_file)
 
         result = None
         for item in json_data:
-            if file_path in item:
-                result = item.get(file_path, [])
+            if file.file_path.path in item:
+                result = item.get(file.file_path.path, [])
                 break
 
-    return render(request, "report.html", {"result": result, "name": file_name})
+    return render(request, "report.html", {"result": result, "name": file.file_name})
